backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP client
httpx==0.26.0

# Date/time
pytz==2024.1
python-dateutil==2.8.2

# Async file handling
aiofiles==23.2.1

# AI
openai==1.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utils
python-dotenv==1.0.0
